requirements.txt
# Mask R-CNN pipeline (install torch/torchvision via: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124)
opencv-python-headless==4.13.0.92
pycocotools==2.0.11
altair==5.5.0
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
axios==0.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
comm==0.2.3
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lxml==6.0.2
macholib==1.15.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
openai==2.21.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.15.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.1
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
zipp==3.23.0
